tools/pyproject.toml
[project]
name = "multigeiger-mqtt-logger"
version = "1.0.0"
description = "MQTT data logger for MultiGeiger radiation sensor"
requires-python = ">=3.8"
dependencies = [
    "paho-mqtt>=1.6.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
analysis = [
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
]

[project.scripts]
mqtt-logger = "mqtt_logger:main"
query-data = "query_data:main"
